requirements.txt
requests
allure-pytest
curlify
pytest